blogapi/requirements.txt
asgiref==3.8.1
certifi==2025.7.14
charset-normalizer==3.4.2
coreapi==2.3.3
coreschema==0.0.4
dj-rest-auth==7.0.1
Django==5.2.1
django-allauth==65.9.0
django-cors-headers==4.7.0
django-rest-swagger==2.2.0
djangorestframework==3.16.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==25.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
